Added dialogue minigame
Updated input to handle holding their state better.
Updated DrawCards to include a zoom function for smaller words.
Updated scene loading to allow us to input a minigame of sorts. This feature in particular needs a lot of work. The programming is very sloppy and needs to be made cleaner and more abstract.