Updated SDLWrapper to provide a scale
Scaling allows the developer to indicate a Window size without breaking any existing code, similar to how PGE provides a way for users to use a pixel scale.
Added an 'over' function to the mouse, allowing users to simply provide a position and a scale of an object instead of manually shape testing for mouse over events.